backend/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "tailored-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Andy Chen"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = "^0.116.1"
uvicorn = {version = "^0.35.0", extras = ["standard"]}
openai = "^1.100.2"
python-dotenv = "^1.1.1"
tiktoken = "^0.11.0"
numpy = "^2.3.2"
gradio = "^5.43.1"
pinecone = "^8.1.0"
youtube-transcript-api = "^1.2.4"
yt-dlp = "^2026.3.17"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
